(wip) starting to add tests
Unfortunately the way I've been importing functions won't help me during testing. This is why I had to import functions during the test method. If I import at the top level then the patch won't work because by then the function is already established in the module that is using it.