http connection errors (#3)
* Allow releases from any branch

The release setup is meant to help us create a release by creating the release branch.  The names "master" and "release" may be different depending on the configuration set in m.json.

* Fix http issues

connection.getresponse() may throw an exception which was not being caught. The response body is type bytes and may gives issues to the json serializer.