Introduce workflows (#13)
* Introduce workflows - git_flow and m_flow

Things are broken until testing is made for each flow.

* Test Config for each flow

* Refactor git_env tests

These do not depend on the workflow so no need to create different test cases.

* Add release_env tests

These cover the git and m workflows. Focus has been done on making sure we get valild tags for releases and hotfixes.

* Provide next tag for hotfixes

Hotfixes prs will be tagged with the next tag.

* Update env vars

Fixes the current build by properly serializing the enum.

* Add bash script to submit pull requests

* quick fixes

* assert_branch should have empty output

* Customize pr messages

Hotfixes are emphazised and instructions are written on the PR description for the gitflow when the release or hotfix is merged back into the develop branch.

* assert_branch should not print

* Fix logic with release environment

Had false positives telling if a build is for a release.

* Rename release scripts

* Add endRelease script

Meant to merge the prs without having us waiting for them. Useful in the gitflow

* fix variable assignment

* fix the endRelease script